logarithmic volume normalization
Added volume scale normalization to the r1:volume control.  this is active whenever the volume is adjusted and the callback is fired. 
Also added provisions for rx_vol to be calculated and set after each tx/rx cycle. Previously the radio would revert to the default rx_vol int, That was set to 100!