anr, indicators, glg tune
Added TXEQ, DSP, and ANR indicators to top right corner of the spectrum scope.
Removed QRO_CONTROL and its related functions. 
Relocated submenu (MENU) show/hide from 1000,-1000 to the main control surface where the QRO enable/stdby button was located.
Relocated QRO enable/stdby button to submenu. 
Added a modified version of GLG TUNE. the original was setting the radio to 2TONE on start up.. this is not what we want. 
It now starts in the last mode recorded.