Log tx power and swr; ADIF export; show power in 1/10 units; comment editing
I changed the data/create_db.sql script, but of course I didn't want to
start over with a fresh database; so I did

  $ sqlite3 ~/sbitx/data/sbitx.db
  sqlite> alter table logbook add column tx_power text default "";
  sqlite> alter table logbook add column vswr text default "";

That puts the columns in the "wrong" order: they come last, but I'd
rather have the comment field last.  Verify it like this:

  sqlite> .schema logbook

logbook_add() detects whether or not the new columns have actually
been added, and omits them if not.

There's no ADIF field defined for SWR, AFAICT.

SWR and power are not editable in the edit dialog so far, but at least
they show up in the list.

But the code should not depend on column order: i.e. it should not do a
query like 'select * from logbook ...' and then expect the columns to be
in a particular order, for display purposes or otherwise. And there is
a lot of room for improvement in export_adif() still.