Replace FONT_ defines with style enum; replace all usages
At least conceptually, in UIs in general, a semantic (such as
"my callsign") applied to a range of text should usually map to a named
style. It's not that we necessarily want to do anything more than change
the font, in practice; but for external UIs, we should export the text
together with semantic metadata, telling what each span of text _means_,
not how the SDR backend expects it to look (that's not the backend's
job to decide). And it's an enum now.