Improve path construction for reading settings
It's a file path, not a directory. And it's wasteful to keep it around
longer than we need it, so put it in a block.

If one searches for "getenv", probably most of those places could
have better implementation of constructing path strings, but most are
not as egregious as this one.