Add write_console_semantic; decruft the console implementation; const
At this point we are back to a one-style-per-line console, but the
foundation is in place to start handling semantic-tagged spans from
ft8_lib rather than guessing about it, and to export semantic-tagged
spans instead of using the ad-hoc markup language.

draw_text() returns length in pixels: it's useful to avoid having to
call measure_text() separately, which ought to reduce the total number
of cairo_x() function calls in the case that it's needed (and otherwise
may be slightly wasteful). But not drawing characters one-by-one ought
to be a bigger win.

Also rearrange styles to group console semantics together, and field
styles below.