FT8: start using ftx_message_offsets_t; restore FT8 color to the console
Measure console_width only once (optimization): it should save some CPU
cycles, at the expense of not being as resizable. (But maybe we shouldn't
be pre-wrapping the text into lines anyway: wrap it on the fly during
drawing?)

Distinguish caller and callee: so far I did not teach ft8lib to
distinguish them; but perhaps it's possible: that would be handy for
supporting multi-part messages later on. But in the meantime, we have to
count them out from the total number of callsigns found. Anyway, IMO
it's nice to show the callee in a dimmer color if it's not my callsign.

Time, SNR, message RST and frequency are all tagged with spans too.
But the first span is for the whole line (STYLE_FT8_RX).