Fix FT8 state machine: hashed calls, prefixes, special CQ; store sent hashes
The hashtable in my sbitx doesn't even contain the hash of my own callsign
unless I let ftx_message_encode() save it (when an ftx_ function takes hash_if,
it should be provided, not NULL!) And if I cold-call a callsign, I need to
store his hash to be ready to handle his replies, too. This becomes quite
noticeable when either callsign is nonstandard and needs to be sent hashed.

So far we still show <> brackets around hashed callsigns in the console,
just to be able to see that it's hashed; but for running the state machine
the <> brackets need to be removed; so add a tokncpy() function to copy
from messages into the m1..m4 variables (which then also has to deal with
"tokens" like R+12, R-20, and callsigns with slashes).

Likewise, strip brackets when replying to messages containing hashed calls.

modem_ft8: override log level

This ft8_lib update also supports messages like "CQ xxxx MYCALL AA01"
or CQ nnn (3 digits, or 4 letters).

FT8 cold call: if it's the first time we've seen this caller, and he has
a complex callsign so that it comes in as a type 4 message, the hash may
not get decoded: i.e. we receive "MYCALL <...> AA01". Replying with
" MYCALL AB02" would be useless: don't do that.