Use ftx_message_encode_[non]std; gen audio @ pitch right before xmit
Whenever we are starting a QSO, pull call_to from the CALL field,
and extra from the my grid.

When auto-replying (executing each state in the state machine), likewise
each of the 3 fields gets populated more directly, rather than
generating message text first and then having ftx_message_encode() pull
it apart again.

Separate sbitx_ftx_msg_audio() from the encode functions, and call it
just-in-time in ft8_start_tx(). Before, it was already spending the CPU
cycles to regenerate audio, but now we do it for the purpose of using
the current TX_PITCH each time. This way it's possible to adjust the
pitch at any time, to get around observed signals, or because of
guessing that if the other station didn't respond, maybe it would make
sense to transmit on a different audio frequency. Before, it was
necessary to hit ESC and then restart the call or the reply, just to use
a different pitch: that's not always possible, depending on how much
traffic has been logged to the console in the meantime.

Don't switch to transmit until/unless there is some audio to transmit
(avoid the tick-tick-tick).