Tapping console line sets X11 selection and calls new ft8_call()
Extract the callsign to call, and the other available data (grid, signal
strength, pitch, time) from the console semantics, rather than
re-parsing the whole line. I.e. we know which callsign we are going to
call because ft8lib told us which range of characters is the "de" call,
not because we just picked the second word in the message and hoped for
the best. Messages like "CQ POTA X1ABC AA01" will not throw it off:
ft8lib knows the callsign is X1ABC, not POTA.

Also rule out calling myself.

ft8_call() depends on fields having been set, but that makes behavior
a bit more predictable: the fields can be checked before calling.
Not that the code is making any use of that yet...

The X11 selection is useful in case you want to paste the whole message
into some other program (a text editor for manual logging, to look up
the callsign on QRZ or something). That works great with remote X11 too.