Tapping console line sets X11 selection and calls new ft8_call()
Extract the callsign to call, and the other available data (grid, signal
strength, time) from the console semantics, rather than re-parsing the
whole line. I.e. we know which callsign we are going to call because
ft8lib told us which range of characters is the "de" call, not because
we just picked the second word in the message and hoped for the best.
Messages like "CQ POTA X1ABC AA01" will not throw it off: ft8lib knows
the callsign is X1ABC, not POTA.

ft8_call() depends on fields having been set, but that makes behavior
a bit more predictable: the fields can be checked before calling.
Not that the code is making any use of that yet...

It bothers me that it's too easy to tap the wrong line on the
touchscreen and start making an unintentional call, or even call myself
accidentally. In this version though, we don't change the tap-to-call
functionality, but I rule out calling myself now.