improve cw decoder in modem_cw.c
- decrease sensitivity to small tuning errors
- adjust sliding window usage to denoise input
- add signal detection hysterisis
- comment additions and cleanup

I think the bulk of the cw decoder part of this file has been little changed since version 3. I read through it (with my editor open) several times, and tried to understand and document how things work. It still follows the original concept, but things are nicer now, better organized, comments added or corrected. I hope people are ok with the level of edits and code moving around - for someone just starting to work on it modem_cw.c is in a better starting place now!