improve cw decoder in modem_cw.c
- decrease sensitivity to small tuning errors
- adjust sliding window usage to denoise input
- add signal detection hysterisis
- use additional bins for probabilistic detection
- comment additions and cleanup

I think the bulk of the cw decoder part of this file had been little changed since version 3. It seems to work better with the changes and improvements

I hope people are ok with the level of edits and code moving around - for someone just starting to work on it modem_cw.c is easier to work on now.