Provide interface for logbook querying and ADIF export
A caller of these functions can query the logbook and export ADIF
without doing sqlite-specific things.

In other branches of this code base, there is another way: not just
saving to a file on the SD card, but exporting over a network interface
(a virtual filesystem). That's not integrated on this branch, but we
might as well make the code easier to keep in sync.