Add xOTA fields to settings dialog and logbook; macros
If xota is set, that and xota_loc get logged. ADIF export is already
done. So, change the settings before your IOTA/SOTA/POTA trip, then
you can export a log containing the references/IDs.

The IOTA/SOTA/POTA macros are similar: the first button sends e.g.
"CQ SOTA MYCALL GRID"; the most-essential buttons are the first 5
(because zbitx only provides 5 macro buttons, so we might as well
get used to using mostly those); and the last button allows you
to enter something else in the CALL field and send CQ THAT ...
(for example CQ BOTA on a beach, or CQ nnn with a 3-digit number).

Selection of the xota mode, xota_loc and macro set are independent.
Setting SOTA mode does not select SOTA.mc, but that's easy already.