Add IOTA/SOTA/POTA fields to logbook; export to ADIF, omit empty fields
I'm not 100% sure whether MY_POTA_REF is right for POTA and simply IOTA
is right for IOTA; but MY_SOTA_REF works for uploading SOTA logs.

We now omit empty fields when exporting ADIF. Most QSOs aren't expected
to have the xota fields, and for that matter there was no point in
exporting empty comment fields either. The ID is also omitted:
it's not important for uploading logs.