Implement FTx CQ_alt and xOTA modes by changing ft8_tx1st to ftx_cq
In FT8 or FT4 modes, "CQ alt" means send CQ only on the first slot of a
4-slot period: i.e. in FT8 mode, send CQ only at 0 seconds and skip the
30-second CQ. In FT4 mode, send CQ only at 0/30 sec.

There are various reasons:
- listen for activity in the even slot as well as the odd
- space out the "repeats" to CQ for a longer period of time
- reduce heat
- conserve batteries
- from a receiver viewpoint, less-frequent CQing is probably
  just as noticeable while generating less clutter

"xOTA" mode means send the xota_loc location (reference ID) in
between the CQ messages, as free text:
- SOTA: send e.g. "SO LA/BU-119"
- POTA: send "PO <park ref>"
- IOTA: send "IO <island ref>"
Let's hope all these references are really 10 characters or less,
since "xO " takes 3 of 13 possible free-text characters.

I haven't found evidence of an existing convention like this,
so let's start one. I think it would be best to send the
location/reference "somehow", and the UI was not making this easy.