modem_ft8: common time format; try to decide what will be sent; decruft
That is, if the user hits the CQ button, or selects a message to answer,
decide whether the message should be FT4 or FT8, whether to send in the
even or odd slot, and whether it's time to start right now.

For the xOTA/Alt modes coming later, ftx_would_send() will also decide
which message to send; that's coming in a later patch.

All time stamps are formatted with tenth-of-a-second resolution now, but
the tenth is omitted if zero (which we often expect in FT8 mode).