xOTA failsafe: don't log xOTA if the location ref hasn't been set
If user_settings.ini has "#xota=" with no value, the GUI might treat it
as POTA rather than NONE (that needs fixing). So I sometimes saw log
entries with POTA PEAK/PARK/ISLE: XOTA got to be POTA somehow, and xOTA
Loc was not set. If it has the default text, it's not set, and this is
not a POTA log entry.