If user_settings.ini has an empty string for "#xota", use NONE
Otherwise we would get a "not permitted" error, and it would default to
POTA in the settings dialog.

Alternatively, I could have fixed this by changing set_field() so that
if the argument is empty, it chooses the first option from an option field.
But this would mean that if you manually issue the command "\xota" with
no argument, it would go back to NONE; and likewise for every other option.
I'm not sure yet if we want that. Also don't understand why this problem
doesn't come up with other option fields.