Move define LOG_LEVEL before ft8_lib includes; improve log messages
The LOG macro is provided by ft8_lib (ft8/debug.h), so we need to define
the log level beforehand if we want to change it from the default.

- FT8 modem: reverse arrows in LOG messages

I keep getting confused, which way is transmit and which way is receive?
The antenna jack is on the right side of the radio physically and on the
schematic too, so I guess it's easier to remember if ">> CQ ..." means
the signal is going out to the right, and "<< A1BCD ..." means it's
received. "-> ..." means it's queued up for later transmission.

Let's keep track of the number of repeats remaining, too. (Too bad the
UI doesn't make it clear)