FT8: don't answer a messsage with my callsign if his callsign is invalid
E.g. this can be triggered by cold-calling from another radio that is using
a non-standard call. If we haven't seen a CQ from him/her, we don't have the
hash in the hashtable. So if the incoming message looks to us like
MYCALL <...> AB01  it's not enough information to be able to answer.
There's no point in sending " MYCALL Rxxx".