Add terminate_qso argument to modem_abort
If the user hits esc, the logger fields are cleared, so it makes sense
to also clear ftx_tx_text in the FTx modem: that means at the next
chance, we will not send anything unless we are starting a new QSO; and
in CQRESP mode, we are free to reply to another CQ, because there is
nothing "queued" up to send.

But in some cases, we don't want to clear ftx_tx_text: e.g. if a 73 needs
to be sent, we should finish sending it; or if we were retrying until
the number of repeats are exhausted, keep doing that.