Fix the mouse wheel with dropdown fields; FTX_CQ=EVEN by default
FTX_CQ defaults to EVEN not ON (but is overridden by settings of course).

We can use the mouse wheel to update a dropdown field without opening it;
but the `#band` field is special and needs a hack for now, because the
`fn` callback is not taking care of editing when the dropdown isn't open.