FTx: Send line number to web UI; expect it back with a selectline command
Trying to re-parse FT8 messages in JS is brittle, unwieldy and unnecessary,
so I want to start moving away from needing that; and, I want to get rid
of ft8_process() because it takes only message text without spans, and thus
also needed to re-parse the message. So we can't depend on it for the web UI
callback; instead, we now have a selectline command that takes a line number.
The sbitx console wraps around rather than keeping all lines forever;
console_current_line is just an index, and wraps around to zero fairly
often, so it's not suitable for a line number for the web UI, which
might keep old messages longer. We assign each line a monotonically
increasing 32-bit unsigned number, send it across as an attribute of the
WSJT-X tag, assign it as the ID of the <li> list item, and then the
FT8_message_chosen() handler is able to pull it out and send it back
with the selectline command. Back on sbitx, if that line number is
recent enough to be found, we call ftx_call_or_continue() to proceed
with the QSO, the same as if it was clicked on the GTK UI.