Add ftx_rx_pitch field to hold chosen station pitch in FTx modes
It needed to be separated from rx_pitch, because that is a user setting:
in CW mode, it's the audio frequency the user would prefer to hear, and
one doesn't expect it to change on its own. Whereas ftx_rx_pitch jumps
each time you choose an FT8/FT4 message to respond to, and controls
the position of the green "needle" in those modes.