Remove the tilde; web UI: rewrite FT8_stylize_message_item; restyle
The tilde was traditional, and other UIs such as the web UI and the
zbitx front panel have been using it as a message parsing aid. But it
wasted 2 characters of space, and space is more precious now that we
have tenth-second time resolution and will soon add the (abbreviated)
country name. The answer is to stop looking for the tilde and re-parsing:
just handle all the hash-letter style tags. But they aren't as unique as
the semantic styles yet, so we still have fewer colors in the web UI
compared to the GTK UI. Tuned up the colors to look decent on the
dark-gray background rather than using the GTK colors exactly.

In a subsequent patch to add the new country field, it shows up in a
smaller font, but bold and bright green so that it stands out.
Grids are amber, "new" callers (not in the log recently) are reddish,
my callsign is bright red, as on the GTK UI.