Restyle RRR and RR73 to de-emphasize; don't update SENT from them
RR73 and RRR are grey rather than bright green: I don't care unless it's
for me. It's ok for 73 IMO because sometimes I might click on one of
those rather than waiting for him to CQ again.

Do not update the SENT field from any of those: that was causing
a bug where the logged "Sent" field no longer matched the report that
was actually sent. But so far we do update it whenever we receive a
message that affects the next report that we can send: that IMO should
reflect the most-recent message received from that call, in case he
turned his beam to hear me better or something like that (tell him that
it worked).