Web UI: handle CW mode decorating every character; ignore CW newlines
In CW mode we need to send one character at a time to the UI. It might be
an RX or TX character. So we output lots of HTML spans: one per character.
Perhaps they could be combined when the style has not changed since the
previous one, if it's worth the trouble. So far, we don't.

Ignore '#J\n' and '#K\n' because they would otherwise turn into
<span>\n</span> which looks the same as a space. When you transmit "AGN?"
cw_update() gets called with '#KA' '#J\n' '#KG' '#KN' '#K?'
and we don't want to see a space between A and G.

Also make the CW RX green color brighter.