Clean up logbook memory management
ASAN complains about memory leaks from this code, but at least the leaks
do not grow much when the logbook is repeatedly opened and closed.

It seems that GTK deletes widgets by default when the window containing
them is closed, so we don't need to add any free/destroy calls for this.