Fix the FTx state machine handling of RR73 and 73 yet again
If country lookup occurs on such messages,
strstr(line + line_len - 5, "RR73") will not work, because
the "RR73" is no longer that close to the end of the line.
It's more efficient not to search the whole line for it.
Whereas looking for RR73 in specific spans was also brittle,
as long as we don't have a special semantic for such things
as 73, RRR and RR73 (not even sure what to call those "words").
Anyway, the intention with doing country lookup when "my call"
is found was just to handle the case when somebody answers my
CQ with a grid. So let's do it only when there is actually a grid.