Add trailing newline to SWR console warning
When outputting a few characters to the console (as in CW mode),
a newline is not written, and the line continues until it wraps.
But a full-line console message should end with a newline.
If this is not done, the semantic spans from whatever comes next
(such as an FT8 message) might get applied to this unfinished line too,
which can be colorful but strange.