Don't clear the logger after 73 is sent in manual mode
The intention with a75dc6de22e6e0a412e27303b2ce2dc334d2db32 was to delay
clearing it until after 73 finishes sending in any of the automated modes.
But in manual mode, we should not clear it at all. Manual means manual:
the user needs to take every step, including logging the QSO; so we shouldn't
interfere by clearing the fields.

In any automated mode, ftx_call_or_continue() logs the QSO when it sees
RR73 or RRR. But in manual mode, it returns early and does not do that;
so as a matter of symmetry, we shouldn't call_wipe() in manual mode either.