Restore bigfont without the reformatting
Here's the minimal patch with just the functionality, so we can actually
see the changes, just those lines in all their "glory".

This is not my change, and I don't approve of how it's done. But I need
to be able to rebase my other patches on top of something closer to what
it was before.