Rules UI: change priorities immediately; add button clears fields
You hit the add button first, fill in details of your new rule,
then save it. Whereas before, if you had selected an old rule,
everything would be populated: then it would be problematic to
change priorities immediately while the spinboxes are adjusted,
because that would affect the old rule, and then you'd save the
new rule afterwards and be confused.

Deleting and re-creating rules is still problematic though because
the database is auto-incrementing the IDs, and we store them in
memory as signed 8-bit. You can have max 128 rules then: there
needs to be a way to reuse old rule IDs when old rules are deleted.
Maybe ID is not a good ordering criterion either...