Avoid putting pitch into EXCH field
"just use the same buffer" was dangerous of course.
After abusing it to store the pitch temporarily, make sure it's null,
in case console_extract_semantic(..STYLE_GRID..) did not succeed.
It won't do anything if there is no grid in the message. Then we
could field_set("EXCH", cand_exch) with the pitch, and then we log
from that field to the logbook, and thus end up with pitch in the
column where the grid should be. I didn't reproduce it when I tried,
but it has definitely happened somehow sometimes.